refactor(navigation): split oversized navigation files
MediaAppNavigation.kt (2093 LOC) -> 65 LOC orchestrator. Each domain
(main/library/detail/player/discovery/collections/sync/onboarding/legacy/
debug) is now a NavGraphBuilder extension function under graphs/.
MediaRoutes, route validation, and NotFoundRouteScreen each move to
their own sibling file.

MediaNavigation.kt (1136 LOC) -> 360 LOC. The MediaNavDestinations
catalog moves to MediaNavDestinations.kt (same package); the sidebar
composables move to MediaSidebar.kt.

ResponsiveNavigation.kt (807 LOC) -> 637 LOC. NavigationItem + the
NavigationItems registry move to NavigationItems.kt.
