+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/components/ResponsiveNavigation.kt
@@ -97,179 +97,9 @@ import com.universalmedialibrary.ui.theme.LocalIsAncientArchitect
 /**
  * Represents a navigation item for bottom navigation, navigation rail, and drawer destinations.
  */
-data class NavigationItem(
-    val route: String,
-    val label: String,
-    val icon: @Composable () -> Unit,
-    val selectedIcon: @Composable () -> Unit = icon,
-    val routeMatch: String? = null,
-    val showInBottomNav: Boolean = true,
-    val preferenceId: String = route
-)
 
-/**
- * Canonical list of app destinations consumed by navigation scaffolds.
- */
-object NavigationItems {
-    val items: List<NavigationItem> = listOf(
-        NavigationItem(
-            route = "home",
-            label = "Home",
-            icon = { Icon(PhosphorIcons.House, contentDescription = "Home") },
-            selectedIcon = { Icon(PhosphorIcons.HouseFill, contentDescription = "Home") }
-        ),
-        NavigationItem(
-            route = "library_details/1",
-            label = "Books",
-            icon = { Icon(PhosphorIcons.Book, contentDescription = "Books") },
-            selectedIcon = { Icon(PhosphorIcons.BookFill, contentDescription = "Books") }
-        ),
-        NavigationItem(
-            route = "library_details/3",
-            label = "Comics",
-            icon = { Icon(PhosphorIcons.Books, contentDescription = "Comics") }
-        ),
-        NavigationItem(
-            route = "library_details/2",
-            label = "Audiobooks",
-            icon = { Icon(PhosphorIcons.Headphones, contentDescription = "Audiobooks") }
-        ),
-        NavigationItem(
-            route = "free_audiobooks",
-            label = "Free Audiobooks",
-            icon = { Icon(PhosphorIcons.Headphones, contentDescription = "Free Audiobooks") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "music",
-            label = "Music",
-            icon = { Icon(PhosphorIcons.MusicNote, contentDescription = "Music") },
-            selectedIcon = { Icon(PhosphorIcons.MusicNoteFill, contentDescription = "Music") }
-        ),
-        NavigationItem(
-            route = "free_music",
-            label = "Free Music",
-            icon = { Icon(Icons.Filled.LibraryMusic, contentDescription = "Free Music") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "hivefy_music",
-            label = "Hivefy",
-            icon = { Icon(Icons.Filled.GraphicEq, contentDescription = "Hivefy") },
-            showInBottomNav = true
-        ),
-        NavigationItem(
-            route = "library_details/4",
-            label = "Movies",
-            icon = { Icon(PhosphorIcons.FilmSlate, contentDescription = "Movies") }
-        ),
-        NavigationItem(
-            route = "free_media",
-            label = "Free Media",
-            icon = { Icon(Icons.Filled.Collections, contentDescription = "Free Media") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "library_details/5",
-            label = "TV Shows",
-            icon = { Icon(PhosphorIcons.Television, contentDescription = "TV Shows") }
-        ),
-        NavigationItem(
-            route = "radio",
-            label = "Radio",
-            icon = { Icon(PhosphorIcons.Radio, contentDescription = "Radio") }
-        ),
-        NavigationItem(
-            route = "visualizer",
-            label = "Visualizer",
-            icon = { Icon(PhosphorIcons.Equalizer, contentDescription = "Visualizer") }
-        ),
-        NavigationItem(
-            route = "ambient",
-            label = "Ambient",
-            icon = { Icon(PhosphorIcons.Nature, contentDescription = "Ambient") }
-        ),
-        NavigationItem(
-            route = "webfiction_manager",
-            label = "Web Fiction",
-            icon = { Icon(Icons.Filled.Language, contentDescription = "Web Fiction") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "story_manager",
-            label = "Story Manager",
-            icon = { Icon(Icons.AutoMirrored.Filled.List, contentDescription = "Story Manager") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "podcasts",
-            label = "Podcasts",
-            icon = { Icon(PhosphorIcons.Microphone, contentDescription = "Podcasts") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "library_details/7",
-            label = "Documents",
-            icon = { Icon(PhosphorIcons.FileText, contentDescription = "Documents") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "collections",
-            label = "Collections",
-            icon = { Icon(PhosphorIcons.Stack, contentDescription = "Collections") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "fanfiction_download",
-            label = "Downloads",
-            icon = { Icon(Icons.Filled.Download, contentDescription = "Downloads") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "storage_browser",
-            label = "Storage",
-            icon = { Icon(Icons.Filled.Storage, contentDescription = "Storage") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "opds_catalog",
-            label = "OPDS",
-            icon = { Icon(Icons.AutoMirrored.Filled.LibraryBooks, contentDescription = "OPDS") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "reading_statistics",
-            label = "Statistics",
-            icon = { Icon(Icons.Filled.BarChart, contentDescription = "Statistics") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "enhanced_search",
-            label = "Search",
-            icon = { Icon(PhosphorIcons.MagnifyingGlass, contentDescription = "Search") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "tag_explorer",
-            label = "Tag Explorer",
-            icon = { Icon(Icons.Filled.Sell, contentDescription = "Tag Explorer") },
-            showInBottomNav = false
-        ),
-        NavigationItem(
-            route = "settings",
-            label = "Settings",
-            icon = { Icon(PhosphorIcons.Gear, contentDescription = "Settings") },
-            selectedIcon = { Icon(PhosphorIcons.GearFill, contentDescription = "Settings") },
-            showInBottomNav = false
-        )
-    )
+// NavigationItem + NavigationItems live in NavigationItems.kt (same package).
 
-    val bottomNavItems: List<NavigationItem> = items.filter { it.showInBottomNav }
-    val settingsItem: NavigationItem = items.first { it.route == "settings" }
-    
-    /** All navigation items available for drawer, including those hidden from bottom bar */
-    val drawerItems: List<NavigationItem> = items
-}
 
 /**
  * Standard bottom navigation bar used on compact width devices.
