+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/navigation/MediaNavigation.kt
@@ -33,786 +33,10 @@ import com.universalmedialibrary.data.settings.BottomBarPreferences
 import com.universalmedialibrary.data.settings.BottomGearPosition
 import com.universalmedialibrary.ui.media.theme.*
 
-/**
- * Clean Media-Centric Navigation System
- * 
- * A beautiful, collapsible sidebar navigation inspired by premium media apps's design.
- * Features:
- * - Collapsible sidebar (full width ↔ icon-only)
- * - Sections with dividers
- * - Hover states and animations
- * - User profile section
- * - Quick actions
- * - Media type badges with colors
- */
 
-// =============================================================================
-// NAVIGATION DATA MODEL
-// =============================================================================
+// MediaNavDestination/NavSection/NavBadge data classes live in MediaNavDestinations.kt (same package).
+// Sidebar composables live in MediaSidebar.kt (same package).
 
-/**
- * Navigation destination
- */
-data class MediaNavDestination(
-    val id: String,
-    val label: String,
-    val icon: ImageVector,
-    val selectedIcon: ImageVector = icon,
-    val route: String,
-    val section: NavSection = NavSection.LIBRARY,
-    val badge: NavBadge? = null,
-    val mediaTypeColor: Color? = null,
-    val enabled: Boolean = true
-)
-
-/**
- * Navigation sections for grouping
- */
-enum class NavSection(val title: String, val order: Int) {
-    HOME("", 0),
-    LIBRARY("Library", 1),
-    DISCOVER("Discover", 2),
-    DOWNLOADS("Downloads", 3),
-    SETTINGS("Settings", 4)
-}
-
-/**
- * Badge to display on navigation item
- */
-data class NavBadge(
-    val count: Int = 0,
-    val text: String? = null,
-    val color: Color = MediaColors.AccentPrimary
-)
-
-/**
- * Default navigation destinations for CleverFerret
- */
-object MediaNavDestinations {
-    
-    val home = MediaNavDestination(
-        id = "home",
-        label = "Home",
-        icon = Icons.Outlined.Home,
-        selectedIcon = Icons.Filled.Home,
-        route = MediaRoutes.HOME,
-        section = NavSection.HOME
-    )
-
-    val search = MediaNavDestination(
-        id = "search",
-        label = "Search",
-        icon = Icons.Outlined.Search,
-        selectedIcon = Icons.Filled.Search,
-        route = MediaRoutes.SEARCH,
-        section = NavSection.HOME
-    )
-
-    val discover = MediaNavDestination(
-        id = "discover",
-        label = "Discover",
-        icon = Icons.Outlined.Explore,
-        selectedIcon = Icons.Filled.Explore,
-        route = MediaRoutes.DISCOVER,
-        section = NavSection.DISCOVER
-    )
-
-    val library = MediaNavDestination(
-        id = "library",
-        label = "Library",
-        icon = Icons.Outlined.VideoLibrary,
-        selectedIcon = Icons.Filled.VideoLibrary,
-        route = MediaRoutes.LIBRARY_ROOT,
-        section = NavSection.LIBRARY
-    )
-    
-    // Library Section
-    val books = MediaNavDestination(
-        id = "books",
-        label = "Books",
-        icon = Icons.Outlined.MenuBook,
-        selectedIcon = Icons.Filled.MenuBook,
-        route = MediaRoutes.BOOKS,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.Book
-    )
-    
-    val audiobooks = MediaNavDestination(
-        id = "audiobooks",
-        label = "Audiobooks",
-        icon = Icons.Outlined.Headphones,
-        selectedIcon = Icons.Filled.Headphones,
-        route = MediaRoutes.AUDIOBOOKS,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.Audiobook
-    )
-    
-    val comics = MediaNavDestination(
-        id = "comics",
-        label = "Comics",
-        icon = Icons.Outlined.AutoStories,
-        selectedIcon = Icons.Filled.AutoStories,
-        route = MediaRoutes.COMICS,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.Comic
-    )
-    
-    val movies = MediaNavDestination(
-        id = "movies",
-        label = "Movies",
-        icon = Icons.Outlined.Movie,
-        selectedIcon = Icons.Filled.Movie,
-        route = MediaRoutes.MOVIES,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.Movie
-    )
-    
-    val tvShows = MediaNavDestination(
-        id = "tv_shows",
-        label = "TV Shows",
-        icon = Icons.Outlined.Tv,
-        selectedIcon = Icons.Filled.Tv,
-        route = MediaRoutes.TV_SHOWS,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.TvShow
-    )
-    
-    val music = MediaNavDestination(
-        id = "music",
-        label = "Music",
-        icon = Icons.Outlined.MusicNote,
-        selectedIcon = Icons.Filled.MusicNote,
-        route = MediaRoutes.MUSIC,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.Music
-    )
-    
-    val podcasts = MediaNavDestination(
-        id = "podcasts",
-        label = "Podcasts",
-        icon = Icons.Outlined.Podcasts,
-        selectedIcon = Icons.Filled.Podcasts,
-        route = MediaRoutes.PODCASTS,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.Podcast
-    )
-    
-    val radio = MediaNavDestination(
-        id = "radio",
-        label = "Radio",
-        icon = Icons.Outlined.Radio,
-        selectedIcon = Icons.Filled.Radio,
-        route = MediaRoutes.RADIO,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.Radio
-    )
-    
-    val documents = MediaNavDestination(
-        id = "documents",
-        label = "Documents",
-        icon = Icons.Outlined.Description,
-        selectedIcon = Icons.Filled.Description,
-        route = MediaRoutes.DOCUMENTS,
-        section = NavSection.LIBRARY,
-        mediaTypeColor = MediaColors.MediaTypes.Document
-    )
-    
-    // Discover Section
-    val webFiction = MediaNavDestination(
-        id = "web_fiction",
-        label = "Web Fiction",
-        icon = Icons.Outlined.Language,
-        selectedIcon = Icons.Filled.Language,
-        route = MediaRoutes.WEB_FICTION,
-        section = NavSection.DISCOVER,
-        mediaTypeColor = MediaColors.MediaTypes.Fanfiction
-    )
-    
-    val opds = MediaNavDestination(
-        id = "opds",
-        label = "OPDS Catalogs",
-        icon = Icons.Outlined.CloudDownload,
-        selectedIcon = Icons.Filled.CloudDownload,
-        route = MediaRoutes.OPDS_BROWSER,
-        section = NavSection.DISCOVER
-    )
-    
-    val freeAudiobooks = MediaNavDestination(
-        id = "free_audiobooks",
-        label = "Free Audiobooks",
-        icon = Icons.Outlined.HeadsetMic,
-        selectedIcon = Icons.Filled.HeadsetMic,
-        route = MediaRoutes.DISCOVER,
-        section = NavSection.DISCOVER,
-        enabled = false
-    )
-    
-    val hivefy = MediaNavDestination(
-        id = "hivefy",
-        label = "Hivefy Music",
-        icon = Icons.Outlined.GraphicEq,
-        selectedIcon = Icons.Filled.GraphicEq,
-        route = MediaRoutes.DISCOVER,
-        section = NavSection.DISCOVER,
-        mediaTypeColor = MediaColors.MediaTypes.Music,
-        enabled = false
-    )
-    
-    val ambient = MediaNavDestination(
-        id = "ambient",
-        label = "Ambient Sounds",
-        icon = Icons.Outlined.Spa,
-        selectedIcon = Icons.Filled.Spa,
-        route = MediaRoutes.AMBIENT_SOUNDS,
-        section = NavSection.DISCOVER
-    )
-
-    val visualizer = MediaNavDestination(
-        id = "visualizer",
-        label = "Visualizer",
-        icon = Icons.Outlined.Equalizer,
-        selectedIcon = Icons.Filled.Equalizer,
-        route = MediaRoutes.VISUALIZER,
-        section = NavSection.DISCOVER
-    )
-    
-    val landseek = MediaNavDestination(
-        id = "landseek",
-        label = "Work with Landseek",
-        icon = Icons.Outlined.Public,
-        selectedIcon = Icons.Filled.Public,
-        route = MediaRoutes.LANDSEEK,
-        section = NavSection.DISCOVER
-    )
-    
-    // Downloads Section
-    val downloads = MediaNavDestination(
-        id = "downloads",
-        label = "Downloads",
-        icon = Icons.Outlined.Download,
-        selectedIcon = Icons.Filled.Download,
-        route = MediaRoutes.DISCOVER,
-        section = NavSection.DOWNLOADS,
-        enabled = false
-    )
-    
-    val storage = MediaNavDestination(
-        id = "storage",
-        label = "Storage",
-        icon = Icons.Outlined.Storage,
-        selectedIcon = Icons.Filled.Storage,
-        route = MediaRoutes.FILE_BROWSER,
-        section = NavSection.DOWNLOADS
-    )
-
-    val sync = MediaNavDestination(
-        id = "sync",
-        label = "Sync",
-        icon = Icons.Outlined.Sync,
-        selectedIcon = Icons.Filled.Sync,
-        route = MediaRoutes.SYNC,
-        section = NavSection.DOWNLOADS
-    )
-
-    val importExport = MediaNavDestination(
-        id = "import_export",
-        label = "Import/Export",
-        icon = Icons.Outlined.ImportExport,
-        selectedIcon = Icons.Filled.ImportExport,
-        route = MediaRoutes.IMPORT_EXPORT,
-        section = NavSection.DOWNLOADS
-    )
-    
-    // Settings Section
-    val collections = MediaNavDestination(
-        id = "collections",
-        label = "Collections",
-        icon = Icons.Outlined.Collections,
-        selectedIcon = Icons.Filled.Collections,
-        route = MediaRoutes.COLLECTIONS,
-        section = NavSection.SETTINGS
-    )
-    
-    val statistics = MediaNavDestination(
-        id = "statistics",
-        label = "Statistics",
-        icon = Icons.Outlined.BarChart,
-        selectedIcon = Icons.Filled.BarChart,
-        route = MediaRoutes.DISCOVER,
-        section = NavSection.SETTINGS,
-        enabled = false
-    )
-    
-    val settings = MediaNavDestination(
-        id = "settings",
-        label = "Settings",
-        icon = Icons.Outlined.Settings,
-        selectedIcon = Icons.Filled.Settings,
-        route = MediaRoutes.SETTINGS,
-        section = NavSection.SETTINGS
-    )
-    
-    /**
-     * All navigation destinations
-     */
-    val allDestinations = listOf(
-        home,
-        search,
-        library,
-        books, audiobooks, comics, movies, tvShows, music, podcasts, radio, documents,
-        discover, webFiction, opds, freeAudiobooks, hivefy, ambient, visualizer, landseek,
-        downloads, storage, sync, importExport,
-        collections, statistics, settings
-    )
-    
-    /**
-     * Primary destinations shown on bottom bar (mobile)
-     */
-    // On phones the bar is horizontally scrollable, so we can expose all major segments.
-    val primaryDestinations = listOf(
-        home,
-        library,
-        discover,
-        search,
-        webFiction,
-        opds,
-        ambient,
-        visualizer,
-        landseek,
-        collections,
-        sync,
-        importExport
-    )
-    
-    /**
-     * Get destinations grouped by section
-     */
-    fun getDestinationsBySection(): Map<NavSection, List<MediaNavDestination>> {
-        return allDestinations.groupBy { it.section }
-            .toSortedMap(compareBy { it.order })
-    }
-}
-
-// =============================================================================
-// SIDEBAR NAVIGATION
-// =============================================================================
-
-/**
- * Clean media-centric collapsible sidebar
- */
-@Composable
-fun MediaSidebar(
-    currentRoute: String,
-    onNavigate: (String) -> Unit,
-    isExpanded: Boolean,
-    onToggleExpanded: () -> Unit,
-    userAvatarUrl: String? = null,
-    userName: String = "User",
-    modifier: Modifier = Modifier
-) {
-    val sidebarWidth by animateDpAsState(
-        targetValue = if (isExpanded) MediaSizes.SidebarWidth else MediaSizes.SidebarCollapsedWidth,
-        animationSpec = tween(durationMillis = MediaAnimations.Duration.Normal),
-        label = "sidebar_width"
-    )
-    
-    val destinationsBySection = remember { MediaNavDestinations.getDestinationsBySection() }
-    
-    Surface(
-        modifier = modifier
-            .width(sidebarWidth)
-            .fillMaxHeight(),
-        color = MediaColors.Background,
-        tonalElevation = MediaElevation.SM
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(vertical = MediaSpacing.MD)
-        ) {
-            // Logo / Collapse Toggle
-            SidebarHeader(
-                isExpanded = isExpanded,
-                onToggleExpanded = onToggleExpanded
-            )
-            
-            Spacer(modifier = Modifier.height(MediaSpacing.LG))
-            
-            // User Profile
-            SidebarUserProfile(
-                avatarUrl = userAvatarUrl,
-                userName = userName,
-                isExpanded = isExpanded
-            )
-            
-            Spacer(modifier = Modifier.height(MediaSpacing.LG))
-            
-            HorizontalDivider(
-                color = MediaColors.Divider,
-                modifier = Modifier.padding(horizontal = MediaSpacing.MD)
-            )
-            
-            Spacer(modifier = Modifier.height(MediaSpacing.MD))
-            
-            // Navigation Items
-            LazyColumn(
-                modifier = Modifier
-                    .weight(1f)
-                    .fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(MediaSpacing.XXS)
-            ) {
-                destinationsBySection.forEach { (section, destinations) ->
-                    // Section Header (only in expanded mode)
-                    if (section != NavSection.HOME && isExpanded) {
-                        item {
-                            SidebarSectionHeader(
-                                title = section.title,
-                                modifier = Modifier.padding(
-                                    start = MediaSpacing.LG,
-                                    top = MediaSpacing.MD,
-                                    bottom = MediaSpacing.SM
-                                )
-                            )
-                        }
-                    }
-                    
-                    // Destinations in this section
-                    items(destinations) { destination ->
-                        SidebarNavItem(
-                            destination = destination,
-                            isSelected = isDestinationSelected(currentRoute = currentRoute, destinationRoute = destination.route),
-                            isExpanded = isExpanded,
-                            onClick = { onNavigate(destination.route) }
-                        )
-                    }
-                }
-            }
-            
-            Spacer(modifier = Modifier.height(MediaSpacing.MD))
-            
-            // Quick Actions at bottom
-            SidebarQuickActions(
-                isExpanded = isExpanded,
-                onSearch = { /* Navigate to search */ },
-                onNotifications = { /* Show notifications */ }
-            )
-        }
-    }
-}
-
-@Composable
-private fun SidebarHeader(
-    isExpanded: Boolean,
-    onToggleExpanded: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = MediaSpacing.MD),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = if (isExpanded) Arrangement.SpaceBetween else Arrangement.Center
-    ) {
-        if (isExpanded) {
-            // App Logo/Name
-            Text(
-                text = "CleverFerret",
-                style = MediaTypography.TitleSmall,
-                color = MediaColors.AccentPrimary,
-                fontWeight = FontWeight.Bold
-            )
-        }
-        
-        // Collapse/Expand button
-        IconButton(
-            onClick = onToggleExpanded,
-            modifier = Modifier.size(MediaSizes.IconLG)
-        ) {
-            Icon(
-                imageVector = if (isExpanded) Icons.Default.ChevronLeft else Icons.Default.Menu,
-                contentDescription = if (isExpanded) "Collapse sidebar" else "Expand sidebar",
-                tint = MediaColors.TextSecondary
-            )
-        }
-    }
-}
-
-@Composable
-private fun SidebarUserProfile(
-    avatarUrl: String?,
-    userName: String,
-    isExpanded: Boolean
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = MediaSpacing.MD),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = if (isExpanded) Arrangement.Start else Arrangement.Center
-    ) {
-        // Avatar
-        Box(
-            modifier = Modifier
-                .size(if (isExpanded) MediaSizes.IconAvatarExpanded else MediaSizes.IconAvatarCollapsed)
-                .clip(CircleShape)
-                .background(MediaColors.BackgroundSurface),
-            contentAlignment = Alignment.Center
-        ) {
-            if (avatarUrl != null) {
-                AsyncImage(
-                    
-                    model = avatarUrl,
-                    contentDescription = "User avatar",
-                    modifier = Modifier.fillMaxSize(),
-                    contentScale = ContentScale.Crop
-                )
-            } else {
-                Icon(
-                    imageVector = Icons.Default.Person,
-                    contentDescription = "Media image",
-                    tint = MediaColors.TextSecondary,
-                    modifier = Modifier.size(MediaSizes.IconSM)
-                )
-            }
-        }
-        
-        AnimatedVisibility(
-            visible = isExpanded,
-            enter = fadeIn() + expandHorizontally(),
-            exit = fadeOut() + shrinkHorizontally()
-        ) {
-            Column(
-                modifier = Modifier.padding(start = MediaSpacing.SM)
-            ) {
-                Text(
-                    text = userName,
-                    style = MediaTypography.LabelLarge,
-                    color = MediaColors.TextPrimary,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-                Text(
-                    text = "Local Library",
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.TextTertiary
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun SidebarSectionHeader(
-    title: String,
-    modifier: Modifier = Modifier
-) {
-    Text(
-        text = title.uppercase(),
-        style = MediaTypography.LabelSmall,
-        color = MediaColors.TextTertiary,
-        fontWeight = FontWeight.SemiBold,
-        letterSpacing = MediaTypography.LabelSmall.letterSpacing * 1.5f,
-        modifier = modifier
-    )
-}
-
-@Composable
-private fun SidebarNavItem(
-    destination: MediaNavDestination,
-    isSelected: Boolean,
-    isExpanded: Boolean,
-    onClick: () -> Unit
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    val isHovered by interactionSource.collectIsHoveredAsState()
-    val isEnabled = destination.enabled
-    
-    val backgroundColor by animateColorAsState(
-        targetValue = when {
-            isSelected -> MediaColors.AccentPrimary.copy(alpha = 0.15f)
-            isHovered -> MediaColors.BackgroundHover
-            else -> Color.Transparent
-        },
-        animationSpec = tween(durationMillis = MediaAnimations.Duration.Fast),
-        label = "nav_item_bg"
-    )
-    
-    val iconColor by animateColorAsState(
-        targetValue = when {
-            !isEnabled -> MediaColors.TextTertiary
-            isSelected -> MediaColors.AccentPrimary
-            else -> MediaColors.TextSecondary
-        },
-        label = "nav_item_icon"
-    )
-    
-    val textColor by animateColorAsState(
-        targetValue = when {
-            !isEnabled -> MediaColors.TextTertiary
-            isSelected -> MediaColors.TextPrimary
-            else -> MediaColors.TextSecondary
-        },
-        label = "nav_item_text"
-    )
-    
-    val contentPadding = if (isExpanded) {
-        PaddingValues(horizontal = MediaSpacing.MD, vertical = MediaSpacing.SM)
-    } else {
-        PaddingValues(MediaSpacing.SM)
-    }
-    
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = MediaSpacing.SM)
-            .clickable(
-                interactionSource = interactionSource,
-                indication = null,
-                onClick = onClick,
-                enabled = isEnabled
-            ),
-        shape = RoundedCornerShape(MediaCorners.SM),
-        color = backgroundColor
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .alpha(if (isEnabled) 1f else 0.55f)
-                .padding(contentPadding),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = if (isExpanded) Arrangement.Start else Arrangement.Center
-        ) {
-            // Selection indicator
-            if (isSelected && isExpanded) {
-                Box(
-                    modifier = Modifier
-                        .width(MediaSizes.SidebarSelectionIndicatorWidth)
-                        .height(MediaSizes.SidebarSelectionIndicatorHeight)
-                        .background(
-                            MediaColors.AccentPrimary,
-                            RoundedCornerShape(MediaCorners.Full)
-                        )
-                )
-                Spacer(modifier = Modifier.width(MediaSpacing.SM))
-            }
-            
-            // Icon
-            Icon(
-                imageVector = if (isSelected) destination.selectedIcon else destination.icon,
-                contentDescription = destination.label,
-                tint = iconColor,
-                modifier = Modifier.size(MediaSizes.IconMD)
-            )
-            
-            // Label (only in expanded mode)
-            AnimatedVisibility(
-                visible = isExpanded,
-                enter = fadeIn() + expandHorizontally(),
-                exit = fadeOut() + shrinkHorizontally()
-            ) {
-                Row(
-                    modifier = Modifier.padding(start = MediaSpacing.SM),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = destination.label,
-                        style = MediaTypography.BodyMedium,
-                        color = textColor,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                        modifier = Modifier.weight(1f, fill = false)
-                    )
-                    
-                    if (!isEnabled) {
-                        Spacer(modifier = Modifier.width(MediaSpacing.SM))
-                        Text(
-                            text = "Soon",
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextTertiary
-                        )
-                    }
-
-                    // Badge
-                    destination.badge?.let { badge ->
-                        Spacer(modifier = Modifier.width(MediaSpacing.SM))
-                        NavItemBadge(badge = badge)
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun NavItemBadge(badge: NavBadge) {
-    Surface(
-        shape = RoundedCornerShape(MediaCorners.Full),
-        color = badge.color.copy(alpha = 0.2f)
-    ) {
-        Text(
-            text = badge.text ?: badge.count.toString(),
-            style = MediaTypography.Badge,
-            color = badge.color,
-            modifier = Modifier.padding(horizontal = MediaSpacing.SM, vertical = MediaSpacing.XXS)
-        )
-    }
-}
-
-@Composable
-private fun SidebarQuickActions(
-    isExpanded: Boolean,
-    onSearch: () -> Unit,
-    onNotifications: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = MediaSpacing.MD),
-        horizontalArrangement = if (isExpanded) Arrangement.SpaceEvenly else Arrangement.Center
-    ) {
-        QuickActionButton(
-            icon = Icons.Default.Search,
-            label = "Search",
-            showLabel = isExpanded,
-            onClick = onSearch
-        )
-        
-        if (isExpanded) {
-            QuickActionButton(
-                icon = Icons.Default.Notifications,
-                label = "Notifications",
-                showLabel = isExpanded,
-                onClick = onNotifications
-            )
-        }
-    }
-}
-
-@Composable
-private fun QuickActionButton(
-    icon: ImageVector,
-    label: String,
-    showLabel: Boolean,
-    onClick: () -> Unit
-) {
-    IconButton(
-        onClick = onClick,
-        modifier = Modifier.size(MediaSizes.IconAvatarExpanded)
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = label,
-            tint = MediaColors.TextSecondary
-        )
-    }
-}
-
-// =============================================================================
-// BOTTOM NAVIGATION (Mobile)
-// =============================================================================
-
-/**
- * Clean media-centric bottom navigation for mobile
- */
 @Composable
 fun MediaBottomNavigation(
     currentRoute: String,
